+++ backend/api/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from teemo.models import Customer, Store
+from api.models import Customer, Store
 # Register your models here.
 admin.site.register(Customer)
 admin.site.register(Store)
