--- backend/teemo/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class TeemoConfig(AppConfig):
-    name = 'teemo'
